pyproject.toml
[tool.poetry]
name = "bazel2snyk"
version = "0.1.0"
description = ""
authors = ["Scott Esbrandt <scott.esbrandt@snyk.io>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "~3.9"
pysnyk = "^0.9.2"
typer = "^0.4.1"

[tool.poetry.dev-dependencies]
# pytest = "^7.1.1"
# black = "^22.3.0"
# pytest-black = "^0.3.12"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
